refactor(flow): move evaluateGate into postgres-approval-gate façade and reduce gate-service.ts to thin shell (BR32 Lot 4 Slice 1 Step 2)
